pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fedramp-ksi"
version = "1.0.0"
description = "FedRAMP 20x KSI-MLA-05 GitHub Action and App for Terraform IaC evaluation"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FedRAMP KSI Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "python-hcl2>=4.3.0",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
app = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.26.0",
    "PyJWT>=2.8.0",
    "cryptography>=41.0.0",
    "pydantic-settings>=2.1.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
fedramp-action = "action.src.main:main"
fedramp-app = "app.main:run"

[tool.setuptools.packages.find]
where = ["."]
include = ["action*", "app*", "shared*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=action --cov=app --cov-report=term-missing"
